src/requirements.txt
opencv-python
numpy
pyautogui
pyscreeze
pynput
Pillow
keyboard